requirements.txt
crewai
gradio
python-dotenv
sqlalchemy
psycopg2-binary
langchain
langchain-openai
openai
pgvector
fastapi
uvicorn
pydantic
